Began work on the SpawnableObjectsGroup script
This script is designed to allow the user to set spawn rate values for a list of objects and then easily get an item from the list that's randomly picked based on the spawn chance.
The spawn rates will automatically adjust as items are added or removed from the list via a custom property drawer.